refactor(abw): deepen editor assist modules
