+++ apps/ai-blog-writer/apps/backend/tests/test_editor_assist_stop_reason.py
@@ -1,14 +1,22 @@
 """Route tests for the itinerary stop selection-reason composer (ADR 0020)."""
 
+from dataclasses import replace
+
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
 import app.features.editor_assist.routes as editor_assist_routes
+import app.features.editor_assist.itinerary_composition as itinerary_composition
+from app.features.editor_assist.dependencies import get_editor_assist_dependencies
 
 
-def _build_client() -> TestClient:
+def _build_client(*, writer=None) -> TestClient:
     app = FastAPI()
     app.include_router(editor_assist_routes.router)
+    dependencies = get_editor_assist_dependencies()
+    if writer is not None:
+        dependencies = replace(dependencies, invoke_writer=writer)
+    app.dependency_overrides[get_editor_assist_dependencies] = lambda: dependencies
     return TestClient(app)
 
 
@@ -16,7 +24,7 @@ def _writer_returning(text: str):
     class _WriterResult:
         def __init__(self) -> None:
             self.text = text
-            self.model_name = editor_assist_routes.DEFAULT_MODEL
+            self.model_name = itinerary_composition.DEFAULT_MODEL
 
     return lambda **_kwargs: _WriterResult()
 
@@ -28,7 +36,7 @@ def _capturing_writer(text: str, sink: dict):
         class _WriterResult:
             def __init__(self) -> None:
                 self.text = text
-                self.model_name = editor_assist_routes.DEFAULT_MODEL
+                self.model_name = itinerary_composition.DEFAULT_MODEL
 
         return _WriterResult()
 
@@ -52,11 +60,7 @@ def _request_body(**overrides):
 
 def test_stop_reason_refines_rough_note(monkeypatch):
     refined = "A buzzy Barranco rooftop whose sunset views and ceviche make it the natural evening anchor."
-    monkeypatch.setattr(
-        editor_assist_routes, "invoke_writer_model", _writer_returning(refined)
-    )
-
-    client = _build_client()
+    client = _build_client(writer=_writer_returning(refined))
     response = client.post(
         "/editor-assist/compose-itinerary-stop-reason", json=_request_body()
     )
@@ -64,18 +68,12 @@ def test_stop_reason_refines_rough_note(monkeypatch):
     assert response.status_code == 200
     payload = response.json()
     assert payload["reason"] == refined
-    assert payload["model_used"] == editor_assist_routes.DEFAULT_MODEL
+    assert payload["model_used"] == itinerary_composition.DEFAULT_MODEL
 
 
 def test_stop_reason_passes_rough_note_and_context_to_writer(monkeypatch):
     sink: dict = {}
-    monkeypatch.setattr(
-        editor_assist_routes,
-        "invoke_writer_model",
-        _capturing_writer("refined reason", sink),
-    )
-
-    client = _build_client()
+    client = _build_client(writer=_capturing_writer("refined reason", sink))
     response = client.post(
         "/editor-assist/compose-itinerary-stop-reason", json=_request_body()
     )
@@ -91,13 +89,9 @@ def test_stop_reason_passes_rough_note_and_context_to_writer(monkeypatch):
 
 
 def test_stop_reason_strips_generation_fence(monkeypatch):
-    monkeypatch.setattr(
-        editor_assist_routes,
-        "invoke_writer_model",
-        _writer_returning("```\nA clean refined reason.\n```"),
+    client = _build_client(
+        writer=_writer_returning("```\nA clean refined reason.\n```")
     )
-
-    client = _build_client()
     response = client.post(
         "/editor-assist/compose-itinerary-stop-reason", json=_request_body()
     )
@@ -107,11 +101,7 @@ def test_stop_reason_strips_generation_fence(monkeypatch):
 
 
 def test_stop_reason_empty_output_is_502(monkeypatch):
-    monkeypatch.setattr(
-        editor_assist_routes, "invoke_writer_model", _writer_returning("   ")
-    )
-
-    client = _build_client()
+    client = _build_client(writer=_writer_returning("   "))
     response = client.post(
         "/editor-assist/compose-itinerary-stop-reason", json=_request_body()
     )
